category updated

+++ src/service/categotyService.js
@@ -1,4 +1,8 @@
 import Category from "../model/categoryModel.js";
+import mongoose from 'mongoose';
+
+const { ObjectId } = mongoose.Types;
+
 export const add = async (data) => {
   const result = await Category.create(data);
   return result;
@@ -8,13 +12,19 @@ export const getAll = async () => {
   const result = await Category.find({}).populate("items");
   return result;
 };
- 
+
 export const getOne = async (id) => {
+  if (!ObjectId.isValid(id)) {
+    throw new Error("Invalid ID format");
+  }
   const result = await Category.findOne({ _id: id });
   return result;
 };
 
 export const update = async (id, data) => {
+  if (!ObjectId.isValid(id)) {
+    throw new Error("Invalid ID format");
+  }
   const result = await Category.updateOne({ _id: id }, data, {
     new: true,
     runValidators: true,
@@ -23,37 +33,40 @@ export const update = async (id, data) => {
 };
 
 export const createSub = async (categoryID, name) => {
+  if (!ObjectId.isValid(categoryID)) {
+    throw new Error("Invalid ID format");
+  }
   const category = await Category.findOne({ _id: categoryID });
 
   if (!category) {
     throw new Error("Category not found");
   }
 
-  category.subCategorys.push({
-    name: name,
-  });
-
+  category.subCategorys.push({ name: name });
   const result = await category.save();
-
-  return result.subCategorys.filter((item) => item.name === name)[0];
+  return result.subCategorys.find((item) => item.name === name);
 };
+
 export const updateSub = async (categoryID, subCategoryID, name) => {
+  if (!ObjectId.isValid(categoryID) || !ObjectId.isValid(subCategoryID)) {
+    throw new Error("Invalid ID format");
+  }
   const category = await Category.findOne({ _id: categoryID });
 
-  category?.subCategorys?.map((item) => {
+  category?.subCategorys?.forEach((item) => {
     if (item?._id.toString() === subCategoryID) {
       item.name = name;
     }
   });
 
   const result = await category.save();
-
-  return result.subCategorys.filter(
-    (item) => item?._id.toString() === subCategoryID
-  )[0];
+  return result.subCategorys.find((item) => item?._id.toString() === subCategoryID);
 };
 
 export const removeSub = async (categoryID, subCategoryID) => {
+  if (!ObjectId.isValid(categoryID) || !ObjectId.isValid(subCategoryID)) {
+    throw new Error("Invalid ID format");
+  }
   const category = await Category.findOne({ _id: categoryID });
 
   if (category) {
@@ -61,15 +74,16 @@ export const removeSub = async (categoryID, subCategoryID) => {
       (item) => item._id.toString() !== subCategoryID
     );
     await category.save();
-    // return { isDeleted: true };
-    return  category.subCategorys;
-
+    return category.subCategorys;
   } else {
     return null;
   }
 };
 
 export const remove = async (id) => {
+  if (!ObjectId.isValid(id)) {
+    throw new Error("Invalid ID format");
+  }
   const result = await Category.deleteOne({ _id: id });
   return result;
 };
